requirements.txt
numpy
scipy
nibabel
matplotlib
loguru
torchinfo
torchtnt
